requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# AI & Embeddings
openai==1.3.7
qdrant-client==1.7.0
tiktoken==0.5.2

# Payments
stripe==7.4.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Background Jobs & Tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
